remove AddressDiscovery.Script module
