+++ lib/core/src/Cardano/Wallet/Primitive/AddressDiscovery/.#Script.hs
@@ -1,0 +1,1 @@
+pawel@pure.2373:1620741226
